Organize lorebook UI components
