+++ src/features/catalog/lorebooks/components/dialogs/LorebookMakerModal.tsx
@@ -4,20 +4,20 @@
 // ──────────────────────────────────────────────
 import { useState, useCallback, useRef } from "react";
 import { useQueryClient } from "@tanstack/react-query";
-import { Modal } from "../../../../shared/components/ui/Modal";
-import { useConnections } from "../../connections/index";
-import { useLorebooks, useCreateLorebook, lorebookKeys } from "../hooks/use-lorebooks";
-import { useUIStore } from "../../../../shared/stores/ui.store";
+import { Modal } from "../../../../../shared/components/ui/Modal";
+import { useConnections } from "../../../connections/index";
+import { useLorebooks, useCreateLorebook, lorebookKeys } from "../../hooks/use-lorebooks";
+import { useUIStore } from "../../../../../shared/stores/ui.store";
 import { Loader2, Wand2, CheckCircle, AlertCircle, ChevronDown, BookOpen, Plus } from "lucide-react";
 import {
   createLorebookEntrySchema,
   lorebookCategorySchema,
-} from "../../../../engine/contracts/schemas/lorebook.schema";
-import type { Lorebook } from "../../../../engine/contracts/types/lorebook";
-import { ProfessorMariWorkingWindow } from "../../../../shared/components/ui/ProfessorMariWorkingWindow";
-import { generateLorebookMaker } from "../../../../engine/generation/makers";
-import { llmApi } from "../../../../shared/api/llm-api";
-import { storageApi } from "../../../../shared/api/storage-api";
+} from "../../../../../engine/contracts/schemas/lorebook.schema";
+import type { Lorebook } from "../../../../../engine/contracts/types/lorebook";
+import { ProfessorMariWorkingWindow } from "../../../../../shared/components/ui/ProfessorMariWorkingWindow";
+import { generateLorebookMaker } from "../../../../../engine/generation/makers";
+import { llmApi } from "../../../../../shared/api/llm-api";
+import { storageApi } from "../../../../../shared/api/storage-api";
 
 interface Props {
   open: boolean;
