+++ src/features/catalog/lorebooks/components/panel/LorebookRow.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from "react";
 import { BookOpen, Camera, Trash2, UserRound } from "lucide-react";
-import type { Lorebook } from "../../../../engine/contracts/types/lorebook";
-import { resolveManagedLocalAssetUrl } from "../../../../shared/api/local-file-api";
-import { cn } from "../../../../shared/lib/utils";
+import type { Lorebook } from "../../../../../engine/contracts/types/lorebook";
+import { resolveManagedLocalAssetUrl } from "../../../../../shared/api/local-file-api";
+import { cn } from "../../../../../shared/lib/utils";
 import { LOREBOOK_CATEGORY_COLORS, LOREBOOK_PANEL_CATEGORIES } from "./lorebook-panel-config";
 
 export function LorebookRow({
