+++ src/features/catalog/lorebooks/shell.ts
@@ -1,6 +1,6 @@
-export * from "./components/CreateLorebookModal";
-export * from "./components/ImportLorebookModal";
-export * from "./components/LorebookEditor";
-export * from "./components/LorebookKeeperReviewModal";
-export * from "./components/LorebookMakerModal";
-export * from "./components/LorebooksPanel";
+export * from "./components/dialogs/CreateLorebookModal";
+export * from "./components/dialogs/ImportLorebookModal";
+export * from "./components/editor/LorebookEditor";
+export * from "./components/dialogs/LorebookKeeperReviewModal";
+export * from "./components/dialogs/LorebookMakerModal";
+export * from "./components/panel/LorebooksPanel";
